refactor(ui): разделён state source management flow
SourceManagementScreenState оставлен root-координатором presentation destination и derived view state
add feed, create folder и move source state вынесены в отдельные focused файлы
для новых state-граней добавлены targeted tests, а async sync tests стабилизированы ожиданием подписок
