+++ RSSReader/Views/SourceManagement/SourceManagementAddFeedPreviewCommand.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct SourceManagementAddFeedPreviewCommand: Equatable, Sendable {
+    let requestID: UUID
+    let urlString: String
+}
