+++ RSSReaderTests/SourceManagement/SourceManagementCreateFolderStateTests.swift
@@ -1,0 +1,33 @@
+import Foundation
+import Testing
+@testable import RSSReader
+
+@Suite("Source Management / Create Folder State")
+@MainActor
+struct SourceManagementCreateFolderStateTests {
+    @Test
+    func createFolderStateValidatesDuplicateNamesAndBuildsPresentation() {
+        var state = SourceManagementCreateFolderState()
+
+        state.applyAvailableFolders(
+            [
+                SourceManagementFolderSummary(
+                    id: UUID(uuidString: "22222222-2222-2222-2222-222222222222")!,
+                    name: "News",
+                    sortOrder: 0,
+                    feedCount: 3
+                )
+            ],
+            isServiceAvailable: true
+        )
+
+        state.updateNameInput(" news ")
+        #expect(state.validationMessage() == "A folder with this name already exists.")
+
+        state.updateNameInput("Research")
+        let presentation = state.derivedPresentation()
+        #expect(presentation.validationMessage == nil)
+        #expect(presentation.isPrimaryActionEnabled)
+        #expect(presentation.placementDescription == "This folder will be added after 1 existing folders.")
+    }
+}
